got the futures producer working and removed the cluster stuff
