+++ simplesource-command-api/src/main/java/io/simplesource/api/CommandAPI.java
@@ -65,6 +65,11 @@ public interface CommandAPI<K, C> {
             .flatMap(v -> queryCommandResult(commandRequest.commandId, timeout));
     }
 
+    /**
+     * Close the API and free up any resources
+     */
+    void close();
+
     @Value
     class Request<K, C> {
         // the aggregate key the command applies to
