+++ simplesource-command-kafka/src/main/java/io/simplesource/kafka/dsl/AggregateSetBuilder.java
@@ -2,14 +2,8 @@ package io.simplesource.kafka.dsl;
 
 import io.simplesource.api.CommandAPISet;
 import io.simplesource.api.CommandAPI;
-import io.simplesource.kafka.api.RemoteCommandResponseStore;
-import io.simplesource.kafka.internal.KafkaCommandAPI;
-import io.simplesource.kafka.internal.cluster.ClusterSubsystem;
+import io.simplesource.kafka.internal.client.KafkaCommandAPI;
 import io.simplesource.kafka.internal.streams.EventSourcedStreamsApp;
-import io.simplesource.kafka.internal.streams.statestore.CommandResponseStoreBridge;
-import io.simplesource.kafka.internal.streams.statestore.KafkaStreamCommandResponseStoreBridge;
-import io.simplesource.kafka.internal.streams.statestore.KafkaStreamAggregateStoreBridge;
-import io.simplesource.kafka.internal.streams.statestore.AggregateStoreBridge;
 import io.simplesource.kafka.internal.util.NamedThreadFactory;
 import io.simplesource.kafka.spec.AggregateSetSpec;
 import io.simplesource.kafka.spec.AggregateSpec;
@@ -17,8 +11,6 @@ import io.simplesource.kafka.spec.CommandSpec;
 import io.simplesource.kafka.spec.KafkaExecutionSpec;
 import org.apache.kafka.streams.KafkaStreams;
 import org.apache.kafka.streams.KeyValue;
-import org.apache.kafka.streams.StreamsBuilder;
-import org.apache.kafka.streams.StreamsConfig;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -34,24 +26,14 @@ public final class AggregateSetBuilder {
     private Map<String, AggregateSpec<?, ?, ?, ?>> aggregateConfigMap = new HashMap<>();
 
     private static Function<AggregateSpec<?, ?, ?, ?>, KeyValue<String, CommandAPI<?, ?>>> createAggregate(
-            final KafkaConfig kafkaConfig,
-            final KafkaStreams kafkaStreams,
-            final RemoteCommandResponseStore remoteCommandResponseStore,
-            final ScheduledExecutorService scheduledExecutor
+            final KafkaConfig kafkaConfig
     ) {
         return aggregateSpec -> {
             CommandSpec<?, ?> commandSpec = aggregateSpec.getCommandSpec();
-            final CommandResponseStoreBridge commandResponseStoreBridge = new KafkaStreamCommandResponseStoreBridge<>(
-                    commandSpec,
-                    kafkaStreams);
             final CommandAPI commandAPI =
                     new KafkaCommandAPI(
                             commandSpec,
-                    kafkaConfig,
-                    commandResponseStoreBridge,
-                    remoteCommandResponseStore,
-                    scheduledExecutor,
-                    commandSpec.retryDelay());
+                    kafkaConfig);
 
             return KeyValue.pair(commandSpec.aggregateName(), commandAPI);
         };
@@ -100,17 +82,9 @@ public final class AggregateSetBuilder {
         //Hack to get round circular ref
         final CommandAPISet[] aggregatesRef = new CommandAPISet[1];
 
-        final ClusterSubsystem clusterSubsystem =
-                new ClusterSubsystem((aggName) -> aggregatesRef[0].getCommandAPI(aggName), kafkaConfig.clusterConfig(), scheduledExecutor);
-
         final Map<String, CommandAPI<?, ?>> commandApis = aggregateSetSpec.aggregateConfigMap().values()
                 .stream()
-                .map(createAggregate(
-                        aggregateSetSpec.executionSpec().kafkaConfig(),
-                        kafkaStreams,
-                        clusterSubsystem,
-                        aggregateSetSpec.executionSpec().scheduledExecutor()
-                ))
+                .map(createAggregate(aggregateSetSpec.executionSpec().kafkaConfig()))
                 .collect(Collectors.toMap(kv -> kv.key, kv -> kv.value));
 
         CommandAPISet commandAPISet = new CommandAPISet() {
@@ -121,7 +95,6 @@ public final class AggregateSetBuilder {
         };
 
         aggregatesRef[0] = commandAPISet;
-        clusterSubsystem.start();
 
         return commandAPISet;
     }
