--- simplesource-command-kafka/src/main/java/io/simplesource/kafka/internal/cluster/Client.java
@@ -1,10 +1,0 @@
-package io.simplesource.kafka.internal.cluster;
-
-import org.apache.kafka.streams.state.HostInfo;
-
-@FunctionalInterface
-public interface Client {
-
-    void send(HostInfo targetHost, Message message);
-
-}
