--- simplesource-command-kafka/src/main/java/io/simplesource/kafka/internal/cluster/CommandAPILoader.java
@@ -1,9 +1,0 @@
-package io.simplesource.kafka.internal.cluster;
-
-import io.simplesource.api.CommandAPI;
-
-@FunctionalInterface
-public interface CommandAPILoader {
-
-    CommandAPI<?, ?> get(String aggregateName);
-}
